DeathStar built + pictures
Run the last tab in each perspective to print the list of organisations
in the Death Star. The Death Star contains the highest-k K-cores.
K-cores are cohesive structures: maximal connected subgraphs where all
nodes have degree at least K.

+++ H2020/H2020_Code/edgeStacker.py
@@ -62,7 +62,7 @@ def findEdge(node1, node2, graph1, directed = False, create = True):
 
 	
 def main(graph): 
-	projectsTogether =  graph.getIntegerProperty("projectsTogether")
+	projectsTogether =  graph.getDoubleProperty("projectsTogether")
 	moneyTogether = graph.getDoubleProperty('moneyTogether')
 	viewLayout =  graph.getLayoutProperty("viewLayout")
 	ShortName =  graph.getStringProperty("ShortName")
